Update index.hbs
Matt - proposed to change the text to the text from the PDF Nupoor linked in her comment. Looks like there are multiple headers with "about block x block" so I am not sure i proposed adding the text in the correct location. Also, was unclear how to add the data info text (also in the pdf Nupoor linked) with correct formatting so I did not copy/paste that here. Would like to include on the page as planned.

Thank you! LB